Move MagScriptEditor into its own file
