+++ UnityProject/Assets/Editor/MagScriptEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fcfbca67deb3bd64a8d4fd1f842790b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
